refactor: split widgets.py and lagrangian_dynamics.py into focused modules
Closes #218

widgets.py (489 LOC) split into:
  - labelled_slider.py  (LabelledSlider, 72 LOC)
  - parameter_sidebar.py  (ParameterSidebar, 265 LOC)
  - playback_controls.py  (PlaybackControls, 54 LOC)
  - widgets.py is now a 20-LOC re-export shim for backward compat

lagrangian_dynamics.py (463 LOC) decomposed into:
  - lagrangian_kinematics.py  (LagrangianKinematicsMixin with FK/COM/bar methods, 148 LOC)
  - lagrangian_balance.py  (balance_pose and _standing_balanced helpers, 87 LOC)
  - lagrangian_dynamics.py  (core physics class only, 287 LOC)

All backward-compatible imports preserved; 288 tests pass.
